Files in the components folder restructured, every organizational-presentational component pairs are in separate folders, with additional styling file as well
