+++ src/components/ImageUploader/layout.js
@@ -1,36 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useDropzone } from 'react-dropzone';
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(theme => ({
-  container: {
-    marginLeft: '5vw',
-    marginTop: '5vw',
-    width: '40vw',
-    height: '40vw',
-    borderWidth: 2,
-    borderRadius: 2,
-    borderColor: '#eeeeee',
-    borderStyle: 'dashed',
-    backgroundColor: '#fafafa',
-    color: '#bdbdbd',
-    outline: 'none',
-    transition: 'border .24s ease-in-out',
-    alignItems: 'center',
-    display: 'table'
-  },
-  content: {
-    display: 'table-cell',
-    verticalAlign: 'middle'
-  },
-  img: {
-    objectFit: 'contain',
-    width: '100%',
-    height: '100%',
-    minWidth: '100%',
-    minHeight: '100%'
-  }
-}))
+import { useStyles } from './styles'
 
 export default function ImageUploader(props) {
   const [files, setFiles] = useState([]);
